stop/start script, provider update

+++ stop.sh
@@ -2,5 +2,10 @@
 PROJECT=$1
 BUCKET=$2
 
+if [[ $PROJECT == "" || $BUCKET == "" ]]; then
+	echo "Usage: stop.sh <project-id> <staging-bucket>"
+	exit 1
+fi
+
 cd terraform
 terraform destroy -var "project=$PROJECT" -var "bucket=$BUCKET"
